requirements.txt
boto3==1.35.37
botocore==1.35.37
fuzzywuzzy==0.18.0
numpy==2.1.2
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.39.0
streamlit_autorefresh==1.0.1
groq==0.12.0
python-Levenshtein==0.26.1
